[fe#15] refactor: 카드 컴포넌트 스타일 분리
