+++ fe/src/components/Card.js
@@ -1,5 +1,14 @@
-import styled, { css } from 'styled-components';
-import { applyFlex } from '../helper/utils';
+import {
+  CardWrapper,
+  CutPrice,
+  Event,
+  Price,
+  PriceWrapper,
+  StyledH2,
+  StyledImg,
+  StyledP,
+  StyledWrapper,
+} from '../styles/Card';
 
 export function Card({ product }) {
   return (
@@ -33,90 +42,3 @@ function EventWrapper(props) {
     </StyledWrapper>
   );
 }
-
-function styleColorToEvent({ eventName }) {
-  if (eventName === '런칭특가') {
-    return css`
-      width: 76px;
-      height: 30px;
-      background: #ff8e14;
-    `;
-  }
-  if (eventName === '이벤트특가') {
-    return css`
-      width: 87px;
-      height: 30px;
-      background: #6dd028;
-    `;
-  }
-}
-
-const StyledImg = styled.img`
-  width: 411px;
-  height: 411px;
-`;
-
-const CardWrapper = styled.div`
-  margin-right: 24px;
-`;
-
-const StyledH2 = styled.h2`
-  font-family: 'Noto Sans KR';
-  font-style: normal;
-  font-weight: 500;
-  font-size: 16px;
-  line-height: 26px;
-  letter-spacing: -0.008em;
-  color: #3f3f3f;
-  margin-top: 16px;
-`;
-
-const StyledP = styled.p`
-  font-family: 'Noto Sans KR';
-  font-style: normal;
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 24px;
-  letter-spacing: -0.008em;
-  color: #777777;
-  margin: 8px 0px;
-`;
-
-const PriceWrapper = styled.div`
-  ${props => applyFlex(props)}
-  font-family: 'Noto Sans KR';
-  font-style: normal;
-`;
-
-const CutPrice = styled.div`
-  font-weight: 500;
-  font-size: 16px;
-  line-height: 26px;
-  letter-spacing: -0.008em;
-  color: #1b1b1b;
-  margin-right: 8px;
-`;
-
-const Price = styled.div`
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 24px;
-  letter-spacing: -0.008em;
-  text-decoration-line: line-through;
-  color: #bcbcbc;
-`;
-
-const StyledWrapper = styled.div`
-  ${props => applyFlex(props)}
-  margin-top:16px;
-  color: #fff;
-  font-weight: 500;
-  font-size: 12px;
-`;
-
-const Event = styled.div`
-  ${props => applyFlex(props)}
-  ${props => styleColorToEvent(props)}
-  margin-right: 8px;
-  border-radius: 999px;
-`;
